8293331: Refactor FileDispatcherImpl into operating system-specific components
