+++ src/java.base/aix/classes/sun/nio/ch/FileDispatcherImpl.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2000, 2022, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package sun.nio.ch;
+
+import java.io.FileDescriptor;
+import java.io.IOException;
+
+public class FileDispatcherImpl extends UnixFileDispatcherImpl {
+    FileDispatcherImpl() {
+        super();
+    }
+
+    int force(FileDescriptor fd, boolean metaData) throws IOException {
+        return force0(fd, metaData);
+    }
+
+    protected long transferTo(FileDescriptor src, long position,
+                              long count, FileDescriptor dst,
+                              boolean append)
+    {
+        return transferTo0(src, position, count, dst, append);
+    }
+
+    // --- native methods ---
+
+    static native int force0(FileDescriptor fd, boolean metaData)
+        throws IOException;
+
+    private static native long transferTo0(FileDescriptor src, long position,
+                                           long count, FileDescriptor dst,
+                                           boolean append);
+}
