+++ src/java.base/share/classes/sun/nio/ch/FileDispatcher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2007, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2007, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -68,6 +68,22 @@ abstract class FileDispatcher extends NativeDispatcher {
     abstract boolean transferToDirectlyNeedsPositionLock();
 
     abstract boolean canTransferToFromOverlappedMap();
+
+    abstract long allocationGranularity();
+
+    abstract long map(FileDescriptor fd, int prot, long position, long length,
+                      boolean isSync)
+        throws IOException;
+
+    abstract int unmap(long address, long length);
+
+    abstract int maxDirectTransferSize();
+
+    abstract long transferTo(FileDescriptor src, long position, long count,
+                             FileDescriptor dst, boolean append);
+
+    abstract long transferFrom(FileDescriptor src, FileDescriptor dst,
+                               long position, long count, boolean append);
 
     abstract int setDirectIO(FileDescriptor fd, String path);
 }
